add logout, forgot_password, and reset_password

+++ app/crud/base.py
@@ -3,8 +3,9 @@ from fastapi.encoders import jsonable_encoder
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
 from app.core.database import Base
+from datetime import datetime
 
-ModelType = TypeVar("ModelType", bound=Base) # type: ignore
+ModelType = TypeVar("ModelType", bound=Base)
 CreateSchemaType = TypeVar("CreateSchemaType", bound=BaseModel)
 UpdateSchemaType = TypeVar("UpdateSchemaType", bound=BaseModel)
 T = TypeVar('T')
@@ -23,10 +24,24 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         self.model = model
 
     def get(self, db: Session, id: Any) -> Optional[ModelType]:
-        return db.query(self.model).filter(self.model.id == id).first()
+        query = db.query(self.model).filter(self.model.id == id)
+        if hasattr(self.model, 'deleted_at'):
+            query = query.filter(self.model.deleted_at == None)
+        return query.first()
+
+    def get_multi(
+        self, db: Session, *, skip: int = 0, limit: int = 100
+    ) -> List[ModelType]:
+        query = db.query(self.model)
+        if hasattr(self.model, 'deleted_at'):
+            query = query.filter(self.model.deleted_at == None)
+        return query.offset(skip).limit(limit).all()
 
     def get_by_email(self, db: Session, email: str) -> Optional[ModelType]:
-        return db.query(self.model).filter(self.model.email == email).first()
+        query = db.query(self.model).filter(self.model.email == email)
+        if hasattr(self.model, 'deleted_at'):
+            query = query.filter(self.model.deleted_at == None)
+        return query.first()
 
     def create(self, db: Session, *, obj_in: CreateSchemaType, commit: bool = True) -> ModelType:
         obj_in_data = jsonable_encoder(obj_in)
@@ -38,11 +53,7 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         return db_obj
 
     def update(
-        self,
-        db: Session,
-        *,
-        db_obj: ModelType,
-        obj_in: Union[UpdateSchemaType, Dict[str, Any]]
+        self, db: Session, *, db_obj: ModelType, obj_in: Union[UpdateSchemaType, Dict[str, Any]]
     ) -> ModelType:
         obj_data = jsonable_encoder(db_obj)
         if isinstance(obj_in, dict):
@@ -59,6 +70,15 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
 
     def delete(self, db: Session, *, id: int) -> ModelType:
         obj = db.query(self.model).get(id)
-        db.delete(obj)
-        db.commit()
+        if not obj:
+            return None
+
+        if hasattr(self.model, 'deleted_at'):
+            setattr(obj, 'deleted_at', datetime.utcnow())
+            db.add(obj)
+            db.commit()
+            db.refresh(obj)
+        else:
+            db.delete(obj)
+            db.commit()
         return obj
